+++ lib/components/MenuV2/MenuGroupLabel.tsx
@@ -1,0 +1,10 @@
+import * as RadixMenu from "@radix-ui/react-dropdown-menu";
+import styled from "styled-components";
+
+
+const MenuGroupLabel = styled(RadixMenu.Label)(() => ({
+  padding: "6px 8px",
+  userSelect: "none",
+}));
+
+export default MenuGroupLabel;
