+++ lib/components/MenuV2/MenuGroupSeperator.tsx
@@ -1,0 +1,10 @@
+import * as RadixMenu from "@radix-ui/react-dropdown-menu";
+import styled from "styled-components";
+
+const MenuGroupSeperator = styled(RadixMenu.Separator)(() => ({
+  height: 1,
+  backgroundColor: "#eee",
+  margin: "6px 0",
+}));
+
+export default MenuGroupSeperator;
