+++ lib/components/MenuV2/SubMenu.tsx
@@ -1,0 +1,118 @@
+import styled from "styled-components";
+import * as RadixMenu from "@radix-ui/react-dropdown-menu";
+import {
+  MenuItemV2ActionType,
+  MenuItemV2Type,
+  MenuItemV2Variant,
+} from "./types";
+import { FOUNDATION_THEME } from "../../tokens";
+import Block from "../Primitives/Block/Block";
+import Text from "../Text/Text";
+import { contentBaseStyle } from "./MenuV2";
+import Item, { itemBaseStyle } from "./MenuItem";
+import { ChevronRightIcon } from "lucide-react";
+
+export const Sub = styled(RadixMenu.Sub)(() => ({
+  padding: 0,
+  margin: 0,
+  listStyle: "none",
+}));
+
+const SubTrigger = styled(RadixMenu.SubTrigger)(() => ({
+  ...itemBaseStyle,
+}));
+
+const SubContent = styled(RadixMenu.SubContent)(() => ({
+  ...contentBaseStyle,
+}));
+
+export const SubMenu = ({
+  item,
+  idx,
+}: {
+  item: MenuItemV2Type;
+  idx: number;
+}) => {
+  return (
+    <Sub key={idx}>
+      <SubTrigger asChild>
+        <Block
+          padding="6px"
+          display="flex"
+          flexDirection="column"
+          gap={4}
+          width="100%"
+        >
+          <Block display="flex" alignItems="center" gap={4} width="100%">
+            {item.slot1 && (
+              <Block flexShrink={0} height="auto" contentCentered>
+                {item.slot1}
+              </Block>
+            )}
+            <Block
+              display="flex"
+              flexGrow={1}
+              alignItems="center"
+              maxWidth="100%"
+              overflow="hidden"
+            >
+              <Text
+                variant="body.md"
+                color={
+                  item.variant === MenuItemV2Variant.ACTION
+                    ? item.actionType === MenuItemV2ActionType.PRIMARY ||
+                      !item.actionType
+                      ? FOUNDATION_THEME.colors.primary[600]
+                      : FOUNDATION_THEME.colors.red[600]
+                    : FOUNDATION_THEME.colors.gray[600]
+                }
+                fontWeight={500}
+                truncate
+              >
+                {item.label}
+              </Text>
+            </Block>
+            {item.slot2 && (
+              <Block flexShrink={0} height="auto" contentCentered>
+                {item.slot2}
+              </Block>
+            )}
+            {item.slot3 && (
+              <Block flexShrink={0} height="auto" contentCentered>
+                {item.slot3}
+              </Block>
+            )}
+            {item.slot4 && (
+              <Block flexShrink={0} height="auto" contentCentered>
+                {item.slot4}
+              </Block>
+            )}
+
+            <Block flexShrink={0} size="auto" contentCentered>
+              <ChevronRightIcon size={16} />
+            </Block>
+          </Block>
+          {item.subLabel && (
+            <Block display="flex" alignItems="center" width="100%">
+              <Text
+                variant="body.sm"
+                color={FOUNDATION_THEME.colors.gray[400]}
+                fontWeight={400}
+              >
+                {item.subLabel}
+              </Text>
+            </Block>
+          )}
+        </Block>
+      </SubTrigger>
+      <RadixMenu.Portal>
+        <SubContent avoidCollisions>
+          {item.subMenu &&
+            item.subMenu.map((subItem, subIdx) => (
+              <Item key={subIdx} item={subItem} idx={subIdx} />
+            ))}
+        </SubContent>
+      </RadixMenu.Portal>
+    </Sub>
+  );
+};
